+++ CHANGELOG.md
@@ -6,16 +6,32 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/swisscom/powergrr/compare/v0.11.0...master)
 
+This release **adds new commands for exporting (downloads) the files archive from a flow or
+a hunt**, **adds support in FileFinder flow for collecting locked files
+using raw file access** and the well known _stat_ action type and finally,
+**fixes the artifact upload mechanism required by newer GRR releases**.
+
 ### Added
 
-* Add pathtype parameter to FileFinder flow which allows using TSK (deprecated) 
-  or NTFS to access locked files, used for e.g. registry files or $MFT on Windows.
+* Add new export commands for exporting files archive from a hunt
+    (`Get-GRRHuntExport`) or from a flow (`Get-GRRFlowExport`) and save the
+    output to a file. This downloads the files found by GRR using e.g.
+    FileFinder or ArtifactCollectorFlow flow.
+* Add _stat_ ActionType parameter to FileFinder flow.
+* Add _pathtype_ parameter to FileFinder flow which _allows using TSK (deprecated) 
+  or NTFS to access locked files_, used for e.g. registry files or $MFT on Windows.
+* Add artifact exist check before uploading a new one (`Add-GRRArtifact`) if
+  it already exists in GRR to prevent a 500 error.
+* Add files archive export functionality in support function
+    (`Invoke-GRRRequest`).
 
 ### Changed
 * Update CHANGELOG, docs and external help file
 
-<!--
 ### Fixed
+* Fix add artifact command (`Add-GRRArtifact`) to work with current GRR
+  release. Replace multi part with a JSON upload.
+<!--
 ### Security
 ### Deprecated
 ### Removed